+++ src/Screen/Field.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace Orchid\Screen;
 
+use Orchid\Screen\Traits\CanSee;
 use Orchid\Screen\Contracts\FieldContract;
 use Orchid\Screen\Exceptions\FieldRequiredAttributeException;
 
@@ -27,6 +28,8 @@ use Orchid\Screen\Exceptions\FieldRequiredAttributeException;
  */
 class Field implements FieldContract
 {
+    use CanSee;
+
     /**
      * View template show.
      *
@@ -212,6 +215,10 @@ class Field implements FieldContract
      */
     public function render()
     {
+        if (! $this->display) {
+            return;
+        }
+
         $this->checkRequired();
         $this->translate();
 
@@ -273,8 +280,9 @@ class Field implements FieldContract
     {
         $modifiers = get_class_methods($this);
 
-        collect($this->getAttributes())->only(array_merge($this->universalAttributes,
-            $this->inlineAttributes))->map(function ($item, $key) use ($modifiers) {
+        collect($this->getAttributes())
+            ->only(array_merge($this->universalAttributes, $this->inlineAttributes))
+            ->map(function ($item, $key) use ($modifiers) {
                 $key = title_case($key);
                 $signature = 'modify'.$key;
                 if (in_array($signature, $modifiers, true)) {
@@ -417,7 +425,7 @@ class Field implements FieldContract
     }
 
     /**
-     * @return $this
+     * @return \Orchid\Screen\Field
      */
     public function vertical(): self
     {
@@ -427,7 +435,7 @@ class Field implements FieldContract
     }
 
     /**
-     * @return $this
+     * @return \Orchid\Screen\Field
      */
     public function horizontal(): self
     {